+++ Source/SwiftLintFramework/String+SwiftLint.swift
@@ -26,6 +26,38 @@ extension String {
     func countOfTailingCharactersInSet(characterSet: NSCharacterSet) -> Int {
         return String(reverse(self)).countOfLeadingCharactersInSet(characterSet)
     }
+  
+    func trim(charactersInSet: NSCharacterSet = NSCharacterSet.whitespaceCharacterSet()) -> String {
+      return "".join(self.componentsSeparatedByCharactersInSet(charactersInSet))
+    }
+    
+    func rangesDelimitedBy(start: String, end: String? = nil) -> [NSRange] {
+      var stackDepth = 0
+      var initialLine: Int? = nil
+      let optionalRanges = map(self.lines()) { (line: Line) -> NSRange? in
+        if line.content.trim() == start {
+          if stackDepth == 0 {
+            initialLine = line.index
+          }
+          
+          stackDepth++
+        } else if line.content.trim() == (end ?? start) && stackDepth > 0 {
+          stackDepth--
+          
+          if stackDepth == 0 {
+            let optionalRange = map(initialLine) {
+              NSMakeRange($0, line.index - $0)
+            }
+            
+            initialLine = nil
+            return optionalRange
+          }
+        }
+        return nil
+      }
+      
+      return flatten(optionalRanges)
+    }
 }
 
 extension NSString {
